pyproject.toml
[tool.poetry]
name = "django-referral"
version = "0.1.0"
description = ""
authors = ["FourZd <shogo.ryaskov@gmail.com>"]
readme = "README.md"
packages = [{include = "django_referral"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.4"
djangorestframework = "^3.15.1"
psycopg2-binary = "^2.9.9"
djangorestframework-simplejwt = "^5.3.1"
drf-spectacular = "^0.27.2"
django-phonenumber-field = {extras = ["phonenumbers"], version = "^7.3.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
